+++ packages/gamut/src/Layout/types.ts
@@ -4,7 +4,13 @@ export type GapTypes = 'rowGap' | 'columnGap';
 
 export type GapSizes = 'sm' | 'md' | 'lg' | 'xl';
 
-export type MediaSizes = 'sm' | 'md' | 'lg';
+export type ResponsiveProperty<T> =
+  | T
+  | {
+      sm: T;
+      md?: T;
+      lg?: T;
+    };
 
 export type ContainerElementProps = {
   className?: string;